Exporting tip errors to HDF5 for static Y0
Analysis:
Average error data is plotted but it would
be useful to export the data to allow further
analysis.

Resolution:
The catheter tip errors are exported to HDF5
files, with a file for each algorithm + sequence
combination. Each catheter's errors are stored
as a dataset within the file.

This has only been done for the static tracking
Y0 notebook. It needs to be added to the other
notebooks.