Exporting tip errors to HDF5 for static Y1 & Y2
Analysis:
Average error data is plotted but it would
be useful to export the data to allow further
analysis for the static notebooks.

Resolution:
Catheter tip errors from the static Y1
and Y2 notebooks are now exported to HDF5
files the same way as for Y0, with a file
for each algorithm + sequence combination.
Each catheter's errors are stored
as a dataset within the file.