Renaming processed data directory
Renaming the processed data subdirectory to "preprocessed": later,
we will enable downloading the raw data and generating the localization
output into a "processed" subdirectory.

Updating the notebooks to use the "preprocessed" subdirectory if
one called "processed" does not exist.

Also updated the static Y1 notebook to add in the HDF export reading
example.