Adding a prototype notebook to run & test pipeline
TestRun notebook reconstructs the raw data, in parallel if possible,
verifies the output compared to the precomputed data, and can be
used to call the individual static & dynamic analysis notebooks.

It needs to be completed: downloading the dataset is not yet
implemented (dataset is local), and only one notebook is called.

A metadata folder has been added to help with running this pipeline.