Empty directory needed for processing
The directory for the missing SRI Y2 cath 306 recordings
still needs to be in the raw & preprocessed file structure.
It had somehow been removed from the preprocessed directory,
maybe when the directory was renamed.

It also needs to be in the csv file so the processed file
structure will also contain this directory.