Run analysis with zenodo sandbox download
A draft dataset has been uploaded the zenodo sandbox,
a temporary location for testing, etc.

- zenodo_get 1.6.1 has a bug that causes downloads from the
sandbox to fail and the version containing the fix is not
compatible with python 3.8. Instead a fix has been applied
to a zenodo_get fork on github: https://github.com/labonny/zenodo_get
  - removing the zenodo_get dependency in env.yaml & installing
the hacky fix instead
- downloading the raw data from this sandbox link and running
the notebook pipeline using this data