Printing stat tests used for invivo comparisons
For paired measurement comparisons, paired t-tests or
Wilcoxon signed rank tests are used depending on if
the samples are normally distributed. The test name
is already output into the csv file, but it would be
helpful to also see this in the notebook.

Returning the test name through the test and plot methods
and printing this in the in vivo notebook