One-tailed Wilcoxon testing for dynamic notebook
Adding a plot_error_box_plot_wilcoxon method for the
dynamic experiment to implement a one-tailed Wilcoxon
test (with options for a two-tailed test).

Note the order of categories used for comparison
is opposite to that used in the plot_error_box_plot
method but is more standard. See the comments for this
method.

The plot_error_box_plot method is still in the
code, and is used for the independent test comparing
sequences.