Fixing missing empty dir for static experiment
A recording is missing for one of the catheters at
one position in the Y2 static grid.

An empty directory for this recording ensures the
calculations are still done correctly. This had
been added previously but after a higher level directory
rename, was somehow lost. Re-adding the empty
directory with a .gitkeep file in it so that git tracks it.