Collecting all tip errors, not just mean
Statistical comparisons should use all the tip errors, so
these are now collected in the aggregate data frame.

This requires a new function in the catheter_utils metrics
module "get_bias_array", which has not yet been committed.

Using the count of 0 to remove the invalid rows from the
aggregate data frame, where the recording was missing
(instead of searching by the known combination of fields).

The statistical testing has not been implemented yet