Match permissions to host user at runtime
Modified dockerfile to allow the container to
write to bind mounts on linux.

This uses MatchHostFsOwner to adjust the container
user uid & gid to match the permissions of
the user running the container: this would allow
writing to bind mounts.

To run it, pass it your id and a bind mount:
 docker run --user "$(id -u):$(id -g)" -v /my/host/reports:/reports --rm -p 8888:8888 mr-tracking:x.y jupyter notebook --ip 0.0.0.0 --no-browser

The container can now write to the /reports directory.

Note: not tested on Mac OS or Windows hosts! The host system
permissions are not an issue on Mac to begin with: this addition
may break something.

To Do: check if this breaks binder