Matching sequence samples for comparison in static
Using Wilcoxon signed rank test to compare sequences
in the static notebook.

This requires the samples to match, in pairs.

Since the HM sequence recordings were shorter, we
need to take a subset of the 3P sequence samples:
did this by taking the first n samples from the
3P sequence recording, where n is the number of samples
for the matching HM recording.

There was one recording missing from a catheter at
one grid location from the 3P sequence, so the
paired recording from the HM sequence was omitted
for this comparison.