Adding optional write to mount
Adding code to detect if there is a /reports directory
and set up output of figures & data to this directory
if so. This will only work when using the Dockerfile.local
built image, with the user and group ids passed in through
the command line. Details are in Dockerfile.local