Static nb: splitting up code for heatmap plots
Cropping the bias list earlier in the notebook
would make the full analysis consistent. This would
make the number of samples match across sequences.

To prepare for this, splitting up the data collection
and plotting for the heatmap errors. To do: crop
the biases in between and use the cropped biases for
all later analysis.